requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
reportlab
supabase
requests
django-cors-headers
